+++ python/pathway/io/s3_csv/__init__.py
@@ -2,9 +2,6 @@
 
 from __future__ import annotations
 
-from typing import Any
-
-from pathway.internals.api import PathwayType
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.schema import Schema
 from pathway.internals.table import Table
@@ -25,10 +22,6 @@ def read(
     autocommit_duration_ms: int | None = 1500,
     name: str | None = None,
     debug_data=None,
-    value_columns: list[str] | None = None,
-    id_columns: list[str] | None = None,
-    types: dict[str, PathwayType] | None = None,
-    default_values: dict[str, Any] | None = None,
     **kwargs,
 ) -> Table:
     """Reads a table from one or several objects in Amazon S3 bucket.
@@ -54,16 +47,6 @@ def read(
             logs and monitoring dashboards. Additionally, if persistence is enabled, it
             will be used as the name for the snapshot that stores the connector's progress.
         debug_data: Static data replacing original one when debug mode is active.
-        value_columns: Names of the columns to be extracted from the files. [will be deprecated soon]
-        id_columns: In case the table should have a primary key generated according to
-            a subset of its columns, the set of columns should be specified in this field.
-            Otherwise, the primary key will be generated randomly. [will be deprecated soon]
-        types: Dictionary containing the mapping between the columns and the data
-            types (``pw.Type``) of the values of those columns. This parameter is optional, and if not
-            provided the default type is ``pw.Type.ANY``. [will be deprecated soon]
-        default_values: dictionary containing default values for columns replacing
-            blank entries. The default value of the column must be specified explicitly,
-            otherwise there will be no default value. [will be deprecated soon]
 
     Returns:
         Table: The table read.
@@ -143,10 +126,6 @@ def read(
             "csv",
             schema=schema,
             csv_settings=csv_settings,
-            value_columns=value_columns,
-            primary_key=id_columns,
-            types=types,
-            default_values=default_values,
             _stacklevel=5,
         )