+++ python/pathway/io/sqlite/__init__.py
@@ -41,13 +41,7 @@ def read(
     Returns:
         Table: The table read.
     """
-    schema, api_schema = read_schema(
-        schema=schema,
-        value_columns=None,
-        primary_key=None,
-        types=None,
-        default_values=None,
-    )
+    schema, api_schema = read_schema(schema)
 
     data_storage = api.DataStorage(
         storage_type="sqlite",